% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrd_est.R
\name{mrd_est}
\alias{mrd_est}
\title{Multivariate Regression Discontinuity Estimation}
\usage{
mrd_est(
  formula,
  data,
  subset = NULL,
  cutpoint = NULL,
  bw = NULL,
  front.bw = NA,
  m = 10,
  k = 5,
  kernel = "triangular",
  se.type = "HC1",
  cluster = NULL,
  verbose = FALSE,
  less = FALSE,
  est.cov = FALSE,
  est.itt = FALSE,
  local = 0.15,
  ngrid = 250,
  margin = 0.03,
  boot = NULL,
  method = c("center", "univ", "front"),
  t.design = NULL,
  stop.on.error = TRUE
)
}
\arguments{
\item{formula}{The formula of the MRDD; a symbolic description of the model to be fitted. This is supplied in the
format of \code{y ~ x1 + x2} for a simple sharp MRDD, or \code{y ~ x1 + x2 | c1 + c2}
for a sharp MRDD with two covariates. Fuzzy MRDD may be specified as
\code{y ~ x1 + x2 + z} where \code{x} is the running variable, and 
\code{z} is the endogenous treatment variable. Covariates are then included in the 
same manner as in a sharp MRDD.}

\item{data}{An optional data frame containing the variables in the model. If not found in \code{data},
the variables are taken from \code{environment(formula)}.}

\item{subset}{An optional vector specifying a subset of observations to be used in the fitting process.}

\item{cutpoint}{An optional numeric vector of length 2 containing the cutpoints at which assignment to the treatment is determined. The default is c(0, 0).}

\item{bw}{A vector specifying the bandwidths at which to estimate the RD. 
Possible values are \code{"IK09"}, \code{"IK12"}, and a user-specified non-negative numeric vector specifying the bandwidths at which to estimate the RD.
The default is \code{"IK12"}. If \code{bw} is \code{"IK12"}, the bandwidth is calculated using the Imbens-Kalyanaraman 
2012 method. If \code{bw}  is \code{"IK09"}, the bandwidth is calculated using 
the Imbens-Kalyanaraman 2009 method. Then the RD is estimated
with that bandwidth, half that bandwidth, and twice that bandwidth. 
If only a single value is passed into the function,
the RD will similarly be estimated at that bandwidth, half that bandwidth, 
and twice that bandwidth.}

\item{front.bw}{An optional non-negative numeric vector specifying the bandwidths at which to estimate the RD for each
of three effects models. If \code{NA}, \code{front.bw} will be determined by cross-validation. The default is \code{NA}.}

\item{m}{A non-negative integer specifying the number of uniformly-at-random samples to draw as search candidates for \code{front.bw},
if \code{front.bw} is \code{NA}. The default is 10.}

\item{k}{A non-negative integer specifying the number of folds for cross-validation to determine \code{front.bw},
if \code{front.bw} is \code{NA}. The default is 5.}

\item{kernel}{A string indicating which kernel to use. Options are \code{"triangular"} 
(default and recommended), \code{"rectangular"}, \code{"epanechnikov"}, \code{"quartic"}, 
\code{"triweight"}, \code{"tricube"}, and \code{"cosine"}.}

\item{se.type}{This specifies the robust standard error calculation method to use. Options are,
as in \code{\link{vcovHC}}, \code{"HC3"}, \code{"const"}, \code{"HC"}, \code{"HC0"}, 
\code{"HC1"}, \code{"HC2"}, \code{"HC4"}, \code{"HC4m"}, \code{"HC5"}. This option 
is overridden by \code{cluster}.}

\item{cluster}{An optional vector specifying clusters within which the errors are assumed
to be correlated. This will result in reporting cluster robust SEs. This option overrides
anything specified in \code{se.type}. It is suggested that data with a discrete running 
variable be clustered by each unique value of the running variable (Lee and Card, 2008).}

\item{verbose}{A logical value indicating whether to print additional information to 
the terminal. The default is \code{FALSE}.}

\item{less}{Logical. If \code{TRUE}, return the estimates of linear and optimal, 
instead of linear, quadratic, cubic, optimal, half and double.}

\item{est.cov}{Logical. If \code{TRUE}, the estimates of covariates will be included. Not
applicable if method is \code{"front"}.}

\item{est.itt}{Logical. If \code{TRUE}, the estimates of ITT will be returned. Not
applicable if method is \code{"front"}.}

\item{local}{The range of neighboring points around the cutoff on the 
standardized scale on each assignment variable, which is a positive number.}

\item{ngrid}{The number of non-zero grid points on each assignment variable,
which is also the number of zero grid points on each assignment variable. Value used in 
Wong, Steiner and Cook (2013) is 2500, which may cause long computational time.}

\item{margin}{The range of grid points beyond the minimum and maximum
of sample points on each assignment variable.}

\item{boot}{The number of bootstrap samples to obtain standard error of estimates.}

\item{method}{The method to estimate rd effect. Options are \code{"center"}, 
\code{"univ"}, \code{"front"}.}

\item{t.design}{The treatment option according to design.
The 1st entry is for X1: \code{"g"} means treatment is assigned 
if X1 is greater than its cutoff, \code{"geq"} means treatment is assigned 
if X1 is greater than or equal to its cutoff, \code{"l"} means treatment is assigned 
if X1 is less than its cutoff, \code{"leq"} means treatment is assigned 
if X1 is less than or equal to its cutoff.
The 2nd entry is for X2.}

\item{stop.on.error}{Logical. If \code{TRUE} (the default), removes bootstraps which cause
error in the \code{integrate} function, and resample till the specified number of 
bootstrap samples are acquired.}
}
\value{
\code{mrd_est} returns an object of \link{class} "\code{mrd}".
}
\description{
\code{mrd_est} estimates treatment effects in a multivariate regression discontinuity design (MRDD) with two assignment variables, 
including the frontier average treatment effect (\code{tau_MRD}) 
and frontier-specific effects (\code{tau_R} and \code{tau_M}) simultaneously.
}
\examples{
set.seed(12345)
x1 <- runif(1000, -1, 1)
x2 <- runif(1000, -1, 1)
cov <- rnorm(1000)
y <- 3 + 2 * (x1 >= 0) + 3 * cov + 10 * (x2 >= 0) + rnorm(1000)
# centering
mrd_est(y ~ x1 + x2 | cov, method = "center", t.design = c("geq", "geq"))
# univariate
mrd_est(y ~ x1 + x2 | cov, method = "univ", t.design = c("geq", "geq"))
# frontier
mrd_est(y ~ x1 + x2 | cov, method = "front", t.design = c("geq", "geq"))
}
\references{
Wong, V. C., Steiner, P. M., Cook, T. D. (2013). 
  Analyzing regression-discontinuity designs with multiple assignment variables: 
  A comparative study of four estimation methods.
  Journal of Educational and Behavioral Statistics, 38(2), 107-141.
  \url{https://journals.sagepub.com/doi/10.3102/1076998611432172}.

Imbens, G., Kalyanaraman, K. (2009). 
  Optimal bandwidth choice for the regression discontinuity estimator 
  (Working Paper No. 14726). National Bureau of Economic Research.
  \url{https://www.nber.org/papers/w14726}.

Imbens, G., Kalyanaraman, K. (2012). 
  Optimal bandwidth choice for the regression discontinuity estimator. 
  The Review of Economic Studies, 79(3), 933-959.
  \url{https://academic.oup.com/restud/article/79/3/933/1533189}.

Lee, D. S., Card, D. (2010).
  Regression discontinuity inference with specification error. 
  Journal of Econometrics, 142(2), 655-674. 
  \doi{10.1016/j.jeconom.2007.05.003}.
}
