% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrd_power.R
\name{mrd_power}
\alias{mrd_power}
\title{Power Analysis of Multivariate Regression Discontinuity}
\usage{
mrd_power(
  num.rep = 100,
  sample.size = 100,
  x1.dist = "normal",
  x1.para = c(0, 1),
  x2.dist = "normal",
  x2.para = c(0, 1),
  x1.cut = 0,
  x2.cut = 0,
  x1.fuzzy = c(0, 0),
  x2.fuzzy = c(0, 0),
  x1.design = NULL,
  x2.design = NULL,
  coeff = c(0.1, 0.5, 0.5, 1, rep(0.1, 9)),
  eta.sq = 0.5,
  alpha.list = c(0.001, 0.01, 0.05)
)
}
\arguments{
\item{num.rep}{A non-negative integer specifying the number of repetitions used to calculate the empirical alpha. The default is 100.}

\item{sample.size}{A non-negative integer specifying the number of observations in each sample. The default is 100.}

\item{x1.dist}{A string specifying the distribution of the first assignment variable, \code{x1}.
Options are \code{"normal"} and  \code{"uniform"}. The default is the \code{"normal"} distribution.}

\item{x1.para}{A numeric vector of length 2 specifying parameters of the distribution of the first assignment variable, \code{x1}.
If \code{x1.dist} is \code{"normal"}, then \code{x1.para} includes the
mean and standard deviation of the normal distribution.
If \code{x1.dist} is \code{"uniform"}, then \code{x1.para} includes the 
upper and lower boundaries of the uniform distribution. The default is c(0,1).}

\item{x2.dist}{A string specifying the distribution of the second assignment variable, \code{x2}.
Options are \code{"normal"} and  \code{"uniform"}. The default is the \code{"normal"} distribution.}

\item{x2.para}{A numeric vector of length 2 specifying parameters of the distribution of the second assignment variable, \code{x2}.
If \code{x2.dist} is \code{"normal"}, then \code{x2.para} includes the
mean and standard deviation of the normal distribution.
If \code{x2.dist} is \code{"uniform"}, then \code{x2.para} includes the 
upper and lower boundaries of the uniform distribution. The default is c(0,1).}

\item{x1.cut}{A numeric vector of length 1 containing the cutpoint at which assignment to the treatment is determined for the first assignment variable, \code{x1}. The default is 0.}

\item{x2.cut}{A numeric vector of length 1 containing the cutpoint at which assignment to the treatment is determined for the second assignment variable, \code{x2}. The default is 0.}

\item{x1.fuzzy}{A numeric vector of length 2 specifying the probabilities to be assigned to the control, in terms of the first
assignment variable, \code{x1}, for individuals in the treatment based on the cutoff, 
and to treatment for individuals in the control based on the cutoff.
For a sharp design, both entries are 0. 
For a fuzzy design, the first entry is the probability to be assigned to 
control for individuals above the cutpoint, and the second entry is the 
probability to be assigned to treatment for individuals below the cutpoint.
The default is c(0,0), indicating a sharp design.}

\item{x2.fuzzy}{Probabilities to be assigned to control in terms of the 2nd
assignment variable X2 for individuals in treatment based on cutoff, 
and to treatment for individuals in control based on cutoff.}

\item{x1.design}{A string specifying the treatment option according to design for \code{x1}. Options are
\code{"g"} (treatment is assigned if \code{x1} is greater than its cutoff),
\code{"geq"} (treatment is assigned if \code{x1} is greater than or equal to its cutoff),
\code{"l"} (treatment is assigned if \code{x1} is less than its cutoff),
and \code{"leq"} (treatment is assigned if \code{x1} is less than or equal to its cutoff).}

\item{x2.design}{A string specifying the treatment option according to design for \code{x2}. Options are
\code{"g"} (treatment is assigned if \code{x2} is greater than its cutoff),
\code{"geq"} (treatment is assigned if \code{x2} is greater than or equal to its cutoff),
\code{"l"} (treatment is assigned if \code{x2} is less than its cutoff),
and \code{"leq"} (treatment is assigned if \code{x2} is less than or equal to its cutoff).}

\item{coeff}{A numeric vector specifying coefficients of variables in the linear model to generate data. Coefficients are in the following order:
\itemize{
\item{The 1st entry is the intercept.}
\item{The 2nd entry is the slope of treatment 1, i.e. treatment effect 1.}
}
The 1st entry is the intercept. 
The 2nd entry is the slope of treatment 1, i.e. treatment effect 1.
The 3rd entry is the slope of treatment 2, i.e. treatment effect 2.
The 4th entry is the slope of treatment, i.e. treatment effect.
The 5th entry is the slope of assignment 1.
The 6th entry is the slope of assignment 2.
The 7th entry is the slope of interaction between assignment 1 and assignment 2.
The 8th entry is the slope of interaction between treatment 1 and assignment 1.
The 9th entry is the slope of interaction between treatment 2 and assignment 1.
The 10th entry is the slope of interaction between treatment 1 and assignment 2.
The 11th entry is the slope of interaction between treatment 2 and assignment 2.
The 12th entry is the slope of interaction between treatment 1, assignment 1 and assignment 2.
The 13th entry is the slope of interaction between treatment 2, assignment 1 and assignment 2.}

\item{eta.sq}{Expected partial eta-squared of the linear model with respect to the 
treatment itself. It is used to control the variance of noise in the linear model.}

\item{alpha.list}{List of significance levels used to calculate the empirical alpha.}
}
\value{
\code{mrd_power} returns the results of 6 estimators as a table of \link{class} 
  "\code{mrdp}", including mean, variance, and power of estimate. The function \code{summary}
  is used to obtain and print a summary of the power analysis.
  The 1st \code{Linear} results of the linear regression estimator 
  of combined RD using the centering approach.
  The 2nd \code{Opt} results of the local linear regression estimator
  of combined RD using the centering approach, 
  with the optimal bandwidth in the IK 2012 paper.
  The 3rd \code{Linear} results of the linear regression estimator 
  of separate RD in terms of x1 using the univariate approach.
  The 4th \code{Opt} results of the local linear regression estimator
  of separate RD in terms of x1 using the univariate approach, 
  with the optimal bandwidth in the IK 2012 paper.
  The 5th \code{Linear} results of the linear regression estimator 
  of separate RD in terms of x2 using the univariate approach.
  The 6th \code{Opt} results of the local linear regression estimator
  of separate RD in terms of x2 using the univariate approach, 
  with the optimal bandwidth in the IK 2012 paper.
}
\description{
\code{mrd_power} computes the empirical probability that an MRD is significant,
i.e. the empirical alpha of null hypothesis: MRD = 0
}
\examples{
\dontrun{
mrd_power(x1.design = "l", x2.design = "l")
mrd_power(x1.dist = "uniform", x1.cut = 0.5, x1.design = "l", x2.design = "l")
mrd_power(x1.fuzzy = c(0.1, 0.1), x1.design = "l", x2.design = "l")
}
}
