% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mfrd.R
\name{print.mfrd}
\alias{print.mfrd}
\title{Print the Multivariate Frontier Regression Discontinuity}
\usage{
\method{print}{mfrd}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{An \code{mfrd} object, typically the result of \code{\link{mfrd_est}}.}

\item{digits}{The number of digits to print.}

\item{...}{Additional arguments.}
}
\description{
\code{print.mfrd} prints a very basic summary of the multivariate frontier regression 
discontinuity. It is based on the print.RD function in the rdd package.
}
