% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rd.R
\name{print.rd}
\alias{print.rd}
\title{Print the Regression Discontinuity}
\usage{
\method{print}{rd}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{An \code{rd} object, typically the result of \code{\link{rd_est}}.}

\item{digits}{A non-negative integer specifying the number of digits to print.}

\item{...}{Additional arguments passed to \code{print}.}
}
\description{
\code{print.rd} prints a basic summary of the regression discontinuity.
\code{print.rd} is based on the \code{print.RD} function in the "rdd" package.
}
\references{
Drew Dimmery (2016). rdd: Regression Discontinuity Estimation. R package
   version 0.57. https://CRAN.R-project.org/package=rdd
}
