% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mrd.R
\name{summary.mrd}
\alias{summary.mrd}
\title{Summarize the Multivariate Regression Discontinuity}
\usage{
\method{summary}{mrd}(object, level = 0.95, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object}{An object of class \code{"mrd"}, usually a result of a call to \code{\link{mrd_est}}.}

\item{level}{A numeric value between 0 and 1 specifying the confidence level for confidence intervals. The default is 0.95.}

\item{digits}{A non-negative integer specifying the number of digits to display.}

\item{...}{Additional arguments.}
}
\value{

}
\description{
\code{summary.mrd} is a \code{summary} method for class \code{"mrd"}.
It is based on \code{summary.RD} function in the "rdd" package.
}
\references{
Drew Dimmery (2016). rdd: Regression Discontinuity Estimation. R package
   version 0.57. https://CRAN.R-project.org/package=rdd
}
