% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treat_assign.R
\name{treat_assign}
\alias{treat_assign}
\title{Treatment Assignment for Regression Discontinuity}
\usage{
treat_assign(x, cut = 0, t.design = "l")
}
\arguments{
\item{x}{The vector of assignment variable X.}

\item{cut}{The cutoff of assignment variables X.}

\item{t.design}{The treatment option according to design.
The entry is for X: \code{"g"} means treatment is assigned 
if X is greater than its cutoff, \code{"geq"} means treatment is assigned 
if X is greater than or equal to its cutoff, \code{"l"} means treatment is assigned 
if X is less than its cutoff, \code{"leq"} means treatment is assigned 
if X is less than or equal to its cutoff.}
}
\value{
\code{treat_assign} returns the treatment variable as a vector according to the design,
  where 1 means the treated group, and 0 means the control group.
}
\description{
\code{treat_assign} computes the treatment variable T based on the cutoff of 
assignment variables X.
This is an internal function and is typically not directly invoked by the user. 
It can be accessed using the triple colon, as in rddapp:::treat_assign().
}
