% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr_check.R
\name{attr_check}
\alias{attr_check}
\title{Attrition Checks}
\usage{
attr_check(x1, y, t, x2 = NULL)
}
\arguments{
\item{x1}{A numeric object containing the assignment variable.}

\item{y}{A numeric object containing the outcome variable, with the same dimensionality 
as \code{x1}.}

\item{t}{A numeric object containing the treatment variable, with the same dimensionality 
as \code{x1} and \code{y}. The treatment variable should be coded as 0 for untreated and 1 for treated.}

\item{x2}{A numeric object containing the secondary assignment variable.}
}
\value{
\code{attr_check} returns a list containing the amount and percentage of missing data for all variables and subgroups, by treatment.
}
\description{
\code{attr_check} reports missing data on treatment variable, assignment variable, and outcome.
Currently, it only supports designs with one assignment variable.
This is an internal function and is typically not directly invoked by the user. 
It can be accessed using the triple colon, as in rddapp:::attr_check().
}
