% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfrd_est.R
\name{mfrd_est}
\alias{mfrd_est}
\title{Multivariate Frontier Regression Discontinuity Estimation}
\usage{
mfrd_est(
  y,
  x1,
  x2,
  c1,
  c2,
  t.design = NULL,
  local = 0.15,
  front.bw = NA,
  m = 10,
  k = 5,
  kernel = "triangular",
  ngrid = 250,
  margin = 0.03,
  boot = NULL,
  cluster = NULL,
  stop.on.error = TRUE
)
}
\arguments{
\item{y}{A numeric object containing outcome variable.}

\item{x1}{A numeric object containing the first assignment variable.}

\item{x2}{A numeric object containing the second assignment variable.}

\item{c1}{A numeric value containing the cutpoint at which assignment to the treatment is determined for \code{x1}.}

\item{c2}{A numeric value containing the cutpoint at which assignment to the treatment is determined for \code{x2}.}

\item{t.design}{A character vector of length 2 specifying the treatment option according to design.
The first entry is for \code{x1} and the second entry is for \code{x2}. Options are  
\code{"g"} (treatment is assigned if \code{x1} is greater than its cutoff),
\code{"geq"} (treatment is assigned if \code{x1} is greater than or equal to its cutoff),
\code{"l"} (treatment is assigned if \code{x1} is less than its cutoff),
and \code{"leq"} (treatment is assigned if \code{x1} is less than or equal to its cutoff).
The same options are available for \code{x2}.}

\item{local}{A non-negative numeric value specifying the range of neighboring points around the cutoff on the 
standardized scale, for each assignment variable. The default is 0.15.}

\item{front.bw}{A non-negative numeric vector specifying the bandwidths at which to estimate the RD for each
of three effects models. If \code{NA}, \code{front.bw} will be determined by cross-validation. The default is \code{NA}.}

\item{m}{A non-negative integer specifying the number of uniformly-at-random samples to draw as search candidates for \code{front.bw},
if \code{front.bw} is \code{NA}. The default is 10.}

\item{k}{A non-negative integer specifying the number of folds for cross-validation to determine \code{front.bw},
if \code{front.bw} is \code{NA}. The default is 5.}

\item{kernel}{A string indicating which kernel to use. Options are \code{"triangular"} 
(default and recommended), \code{"rectangular"}, \code{"epanechnikov"}, \code{"quartic"}, 
\code{"triweight"}, \code{"tricube"}, and \code{"cosine"}.}

\item{ngrid}{A non-negative integer specifying the number of non-zero grid points on each assignment variable,
which is also the number of zero grid points on each assignment variable. The default is 250. The value used in 
Wong, Steiner and Cook (2013) is 2500, which may cause long computational time.}

\item{margin}{A non-negative numeric value specifying the range of grid points beyond the minimum and maximum
of sample points on each assignment variable. The default is 0.03.}

\item{boot}{An optional non-negative integer specifying the number of bootstrap samples to obtain standard error of estimates.}

\item{cluster}{An optional vector specifying clusters within which the errors are assumed
to be correlated. This will result in reporting cluster robust SEs. It is suggested that data with a discrete running 
variable be clustered by each unique value of the running variable (Lee and Card, 2008).}

\item{stop.on.error}{A logical value indicating whether to remove bootstraps which cause error in the \code{integrate} function. If \code{TRUE}, bootstraps which cause error are removed
and resampled until the specified number of 
bootstrap samples are acquired. If \code{FALSE}, bootstraps which cause error are not removed. The default is \code{TRUE}.}
}
\value{
\code{mfrd_est} returns an object of \link{class} "\code{mfrd}".
  The functions \code{summary} and \code{plot} are used to obtain and print a summary and 
  plot of the estimated regression discontinuity. The object of class \code{mfrd} is a list 
  containing the following components:
\item{w}{What is w?}   
\item{est}{Numeric matrix of the estimate of the discontinuity in the outcome under 
  a ... , for each corresponding bandwidth.}
\item{d}{Numeric matrix of the effect size (Cohen's d) for estimate.}
\item{se}{Numeric matrix of the standard error for each corresponding bandwidth and ....}
\item{m_s}{...}
\item{m_h}{...}
\item{m_t}{...}
\item{dat_h}{...}
\item{dat}{A data frame containing...}
\item{obs}{List of the number of observations within the corresponding bandwidth.}
\item{impute}{A logical value indicating whether multiple imputation is used or not.}
\item{call}{The matched call.}
\item{front.bw}{Numeric vector of each bandwidth used in estimation.}
}
\description{
\code{mfrd_est} implements the frontier approach for multivariate regression discontinuity estimation in Wong, Steiner and Cook (2013). 
It is based on the MFRDD code in Stata.
}
\examples{
set.seed(12345)
x1 <- runif(1000, -1, 1)
x2 <- runif(1000, -1, 1)
cov <- rnorm(1000)
y <- 3 + 2 * (x1 >= 0) + 3 * cov + 10 * (x2 >= 0) + rnorm(1000)
mfrd_est(y = y, x1 = x1, x2 = x2, c1 = 0, c2 = 0, t.design = c("geq", "geq"))
}
\references{
Wong, V., Steiner, P, and Cook, T. (2013).
  Analyzing regression discontinuity designs with multiple assignment variables: A comparative study of four estimation methods. 
  Journal of Educational and Behavioral Statistics, 38(2), 107-141. 
  \doi{10.3102/1076998611432172}.

Lee, D. and Card, D. (2008).
  A Regression discontinuity inference with specification error.
  Journal of Econometrics, 142(2), 655-674. 
  \doi{10.1016/j.jeconom.2007.05.003}.
}
