% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rd.R
\name{print.rd}
\alias{print.rd}
\title{Print the Regression Discontinuity}
\usage{
\method{print}{rd}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{An \code{rd} object, typically the result of \code{\link{rd_est}}.}

\item{digits}{The number of digits to print.}

\item{...}{Additional arguments.}
}
\description{
\code{print.rd} prints a very basic summary of the regression discontinuity.
It is based on the print.RD function in the rdd package.
}
