% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_run.R
\name{shiny_run}
\alias{shiny_run}
\title{Launch the Shiny App for Regression Discontinuity}
\usage{
shiny_run(app_name = "shinyrdd")
}
\arguments{
\item{app_name}{The name of Shiny app.}
}
\description{
\code{shiny_run} looks for the Shiny app for rdd and launch it.
}
\examples{
\dontrun{
shiny_run()
shiny_run("shinyrdd")
}
}
